+++ src/main/java/com/zb/deuggeun/common/validation/type/DateOrderValidatable.java
@@ -1,4 +1,4 @@
-package com.zb.deuggeun.common.validation;
+package com.zb.deuggeun.common.validation.type;
 
 import java.time.LocalDate;
